ADD method section ENH abstract with newer viz code
