+++ content/references.md
@@ -1,0 +1,4 @@
+# References 
+
+```{bibliography}
+```
